blogapp/requirements.txt
fastapi>=0.76,<0.78.0
uvicorn>=0.17.1<0.17.6
sqlalchemy>=1.4.30<1.4.37
sqlalchemy-utils>0.37.5<0.38.2
psycopg2>=2.9<2.9.3
python-dotenv>=0.17.0<0.20.0